app/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
pyspellchecker
nltk
python-multipart
jinja2
